+++ src/utils/provider.ts
@@ -1,5 +1,6 @@
 import type { RpcTransactionResponse } from '../rpc/types.gen.js';
 import type { FinalExecutionOutcome } from '../types/index.js';
+import { base64Decode, bytesToString } from './format.js';
 
 /** @hidden */
 // eslint-disable-next-line @typescript-eslint/ban-types
@@ -12,7 +13,7 @@ export function getTransactionLastResult(
         'SuccessValue' in txResult.status &&
         typeof txResult.status.SuccessValue === 'string'
     ) {
-        const value = Buffer.from(txResult.status.SuccessValue, 'base64').toString();
+        const value = bytesToString(base64Decode(txResult.status.SuccessValue));
         try {
             return JSON.parse(value);
         } catch (e) {